refactor: AuthLayout을 src/layout 디렉토리로 이동
