+++ src/App.tsx
@@ -42,7 +42,7 @@ import ShareLocation from "@pages/ShareLocation/ShareLocation";
 import PwaGuide from "@pages/PwaGuide/PwaGuide";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ToastContainer } from "react-toastify";
-import { AuthLayout } from "@components/AuthLayout/AuthLayout";
+import { AuthLayout } from "@/layout/AuthLayout";
 
 // 인증이 불필요한 Public 경로 목록 (백그라운드 복귀 시 토큰 재발급 제외 대상)
 const PUBLIC_PATHS = [
